rename service to be a little nicer
