+++ kahuna/public/js/edits/archiver.html
@@ -1,16 +1,16 @@
 <div class="archiver"
         ng:if="!ctrl.disabled"
-        ng:class="{'archiver--archived': ctrl.isArchived}">
+        ng:class="{'archiver--archived': ctrl.isArchived()}">
 
     <!-- TODO: Start looking at a new loader to replace our "ing…" loader -->
-    <button ng:show="ctrl.isArchived" title="unarchive" ng:click="ctrl.unarchive()">
+    <button ng:show="ctrl.isArchived()" title="unarchive" ng:click="ctrl.unarchive()">
         <gr-icon>star</gr-icon>
         <span class="archiver__label">
             <span ng:show="ctrl.withText">unarchiv<span ng:show="!ctrl.saving">e</span><span ng:show="ctrl.saving">ing…</span></span>
         </span>
     </button>
 
-    <button ng:show="!ctrl.isArchived" title="archive" ng:click="ctrl.archive()">
+    <button ng:show="!ctrl.isArchived()" title="archive" ng:click="ctrl.archive()">
         <gr-icon>star_border</gr-icon>
         <span class="archiver__label">
             <span ng:show="ctrl.withText">archiv<span ng:show="!ctrl.saving">e</span><span ng:show="ctrl.saving">ing…</span></span>
@@ -19,6 +19,6 @@
 </div>
 
 <button class="archiver archiver--archived"
-        ng:if="ctrl.disabled && ctrl.isArchived">
+        ng:if="ctrl.disabled && ctrl.isArchived()">
     <gr-icon>star</gr-icon>
 </button>
