+++ kahuna/public/js/image/stream.js
